requirements.txt
delta-spark==3.1.0
boto3
